+++ src/lib/components/Card/Card.svelte
@@ -1,0 +1,30 @@
+<script lang="ts">
+  import { cn } from "$lib/index.js"
+  import type { AsChild } from "$lib/types.js"
+  import type { HTMLAttributes } from "svelte/elements"
+  import { tv } from "tailwind-variants"
+
+  type $$Props = HTMLAttributes<HTMLDivElement> & AsChild
+
+  export let asChild: $$Props["asChild"] = false
+  export { className as class }
+
+  let className = ""
+
+  const style = tv({
+    base: cn`
+      rounded-xl
+      bg-muted/5
+      border
+      border-muted/5
+    `,
+  })
+</script>
+
+{#if asChild}
+  <slot />
+{:else}
+  <div class={cn(style.base, className)} {...$$restProps}>
+    <slot />
+  </div>
+{/if}
