--- src/lib/components/Card/CardRoot.svelte
@@ -1,22 +1,0 @@
-<script lang="ts">
-  import { cn } from "$lib"
-  import type { HTMLAttributes } from "svelte/elements"
-  import { tv } from "tailwind-variants"
-
-  type $$Props = HTMLAttributes<HTMLDivElement>
-
-  export { className as class }
-
-  let className = ""
-
-  const style = tv({
-    base: cn`
-      rounded-lg
-      bg-muted/5
-    `,
-  })
-</script>
-
-<div class={cn(style.base, className)} {...$$restProps}>
-  <slot />
-</div>
